requirements.txt
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
pyyaml>=6.0
pandas>=2.1
tqdm>=4.66
jupyter>=1.0
